requirements.txt
# Dependencias para el proyecto de reconocimiento de voz y procesamiento de imágenes
numpy>=1.19.0
scipy>=1.5.0
sounddevice>=0.4.0
Pillow>=8.0.0
opencv-python>=4.5.0
torch>=1.9.0
torchvision>=0.10.0